+++ .github/workflows/ci.yml
@@ -466,38 +466,25 @@ jobs:
                   shared-key: ci-${{ runner.os }}
                   save-if: ${{ github.ref == 'refs/heads/main' }}
 
-            - name: Build LiquidCache server
-              run: cargo build --bin example_server
-            - name: Build LiquidCache client
-              run: cargo build --bin example_client
-            - name: Build LiquidCache client projection pushdown
-              run: cargo build --bin example_projection_pushdown
-            - name: Build LiquidCache in process eviction
-              run: cargo build --bin example_inprocess_cache_eviction
-            - name: Build LiquidCache in process insertion
-              run: cargo build --bin example_inprocess_insertion
-            - name: Build LiquidCache in process read
-              run: cargo build --bin example_inprocess_read
+            - name: Build datafusion-local example
+              run: cargo build --bin datafusion-local
+            - name: Build datafusion-client-server example
+              run: cargo build --bin datafusion-client-server
+            - name: Build core example
+              run: cargo build --bin core
 
             - name: Start LiquidCache server
               run: |
-                  env RUST_LOG=info nohup cargo run --bin example_server -- --abort-on-panic &> server.log &
+                  env RUST_LOG=info nohup cargo run --bin datafusion-client-server -- --mode server &> server.log &
                   echo $! > server.pid  # Save PID for later cleanup
                   sleep 2  # Wait for server to start up
 
-            - name: Start LiquidCache client
+            - name: Run datafusion client
               run: |
                   # First run to populate the cache
-                  env RUST_LOG=info cargo run --bin example_client
+                  env RUST_LOG=info cargo run --bin datafusion-client-server -- --mode client
                   # Run twice to test the cache
-                  env RUST_LOG=info cargo run --bin example_client
-
-            - name: Start LiquidCache client projection pushdown
-              run: |
-                  # First run to populate the cache
-                  env RUST_LOG=info cargo run --bin example_projection_pushdown
-                  # Run twice to test the cache
-                  env RUST_LOG=info cargo run --bin example_projection_pushdown
+                  env RUST_LOG=info cargo run --bin datafusion-client-server -- --mode client
 
             - name: Kill LiquidCache server and show logs
               if: always()
@@ -507,17 +494,8 @@ jobs:
                   pkill -F server.pid || true
                   rm -f server.pid
 
-            - name: Start LiquidCache in process projection pushdown
-              run: |
-                  # Run to populate to evict cache
-                  env RUST_LOG=info cargo run --bin example_inprocess_cache_eviction
+            - name: Run datafusion-local example
+              run: env RUST_LOG=info cargo run --bin datafusion-local
 
-            - name: Start LiquidCache in process insert
-              run: |
-                  # Run to populate cache with arrow array
-                  env RUST_LOG=info cargo run --bin example_inprocess_insertion
-
-            - name: Start LiquidCache in process read arrow array
-              run: |
-                  # Run to populate cache and read arrow array
-                  env RUST_LOG=info cargo run --bin example_inprocess_read
+            - name: Run core example
+              run: env RUST_LOG=info cargo run --bin core
