+++ examples/Cargo.toml
@@ -4,35 +4,21 @@ edition = { workspace = true }
 publish = false
 
 [[bin]]
-name = "example_server"
-path = "example_server.rs"
+name = "datafusion-local"
+path = "datafusion-local.rs"
 
 [[bin]]
-name = "example_client"
-path = "example_client.rs"
+name = "datafusion-client-server"
+path = "datafusion-client-server.rs"
 
 [[bin]]
-name = "example_local"
-path = "example_local.rs"
-
-[[bin]]
-name = "example_projection_pushdown"
-path = "example_projection_pushdown.rs"
-
-[[bin]]
-name = "example_inprocess_insertion"
-path = "example_inprocess_insertion.rs"
-
-[[bin]]
-name = "example_inprocess_cache_eviction"
-path = "example_inprocess_cache_eviction.rs"
-
-[[bin]]
-name = "example_inprocess_read"
-path = "example_inprocess_read.rs"
+name = "core"
+path = "core.rs"
 
 [dependencies]
+arrow = { workspace = true }
 datafusion = { workspace = true }
+liquid-cache = { workspace = true }
 liquid-cache-datafusion-server = { workspace = true }
 liquid-cache-datafusion-client = { workspace = true }
 liquid-cache-datafusion-local = { workspace = true }
