+++ examples/datafusion-local.rs
@@ -10,6 +10,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     let (ctx, _) = LiquidCacheLocalBuilder::new()
         .with_max_memory_bytes(1024 * 1024 * 1024) // 1GB
+        .with_max_disk_bytes(1024 * 1024 * 1024 * 10) // 10GB
         .with_cache_dir(temp_dir.path().to_path_buf())
         .with_squeeze_policy(Box::new(TranscodeSqueezeEvict))
         .with_cache_policy(Box::new(LiquidPolicy::new()))
